Updates build config for core package.
I removed the UMD build as it was only added as an experiment for the schema package. The current setup does not produce UMD builds that would work so disabling it makes more sense.

Note that the size-limit tool is not setup properly right now. I'm not sure what the right config would be. I have left it alone for now.